# Improve make translation data performance.