Speed up `maketranslationdata.py` by reducing work done in optimization loop.

|Port| Optimization Setting | Build Time -j16 | Bytes Saved |
| --- | --- | --- | --- |
|RP2 Pico W|1|16.3 sec|8336|
||9 (before this pull)|31.6 sec|12873|
||9 (with this pull)|17.8 sec|13093|
|Atmel-SAMD Pyruler|9 (before this pull)|12.87 sec|4494|
||9 (with this pull)|12.83|4476|

Speed improves, slight variation in compression (improves on RP2, very slightly worse on Atmel-SAMD). Speed on RP2 optimization setting 9 is now close to setting 1, perhaps consider reverting level settings for RP2 port?