@eightycc 2023-10-24
@tannewt SAMD has a smaller pool of strings to compress, so it was already under the size where performance fell off a cliff.